Added an IcicleGraph.js file because the version of react spring that works with TS doens't work with interpolate for zooming. Zooming of the graph is now functional
